Test passing quoted arguments to the commandline utility
If we pass `mongooseimctl` quoted strings then it interprets
characters separated by spaces as other arguments. Essentially, we
want to preserve how the arguments are passed, i.e. invoking with the
single argument "fred bloggs" is not the same as invoking with the two
arguments "fred" "bloggs". Although this only changes one line in one
test case in the group `roster`, it should be enough to verify changes
for any invocation.